requirements.txt
# Стек технологий

numpy<2.0                 # Исправление совместимости с OpenCV
ultralytics==8.0.196      # YOLOv8
torch==2.0.1              # PyTorch (стабильная версия)
torchvision==0.15.2       # Компьютерное зрение для PyTorch
opencv-python==4.8.1.78   # Работа с камерой и изображениями
flask==3.0.0              # Веб-интерфейс (API)
psycopg2-binary==2.9.9    # Работа с PostgreSQL
pyyaml==6.0.1             # Для конфигов
python-dotenv==1.0.0      # Переменная окружения
gunicorn==21.2.0          # Для работы веб-сервера с камерой